机器数据温控校准可以调数

+++ src/main/java/com/laola/apa/mapper/EquipmentStateMapper.java
@@ -63,5 +63,5 @@ public interface EquipmentStateMapper {
      */
     int deleteById(Integer id);
 
-    int temperatureControlCalibration(int i);
+    int temperatureControlCalibration(Float i);
 }
